Commit project submodule
To clearly organize and separate different projects, a submodule to the
course project was added. Documentation on how to clone and update this
repository and its submodules is also included in this commit.

Additionally, project proposal and inspiration gallery were moved to
'hw' folder as this is content related to the class.